update Laws; finish Discussion-Tests; update Collections
updated OgdRisApi and Laws (still not working).
finished testing and debugging Discussions except the sorting by popularity.
updated Collections that had url() missing for the hrefs.
updated Handler to throw CannotResolveDependencyException when foreign keys are not found.
updated ModelFactory slightly to be able to pass a created_at timestamp for some entities.